# PERF: MultiIndex.equals