- [x] closes #43549
- [ ] tests added / passed
- [ ] Ensure all linting tests pass, see [here](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#pre-commit) for how to run them
- [ ] whatsnew entry

```
from dateutil import tz
import pandas as pd
from time import time

dates = pd.date_range('2010-01-01', periods=1000, tz=tz.tzutc())
index = pd.MultiIndex.from_product([range(100), dates])
index2 = index.copy()

%timeit index.equals(index2)
330 ms ± 6.44 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)  # <- master
1.79 ms ± 7.4 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)  # <- PR
```