@jreback 2021-09-16
prob worth backporting this, can you add a whatsnew note @jbrockmendel ping on green