@jreback 2021-09-16
@meeseeksdev backport 1.3.x